track dotfiles in Git
Track .bash_profile. Storing this file allows for a familiar Bash shell
experience without manual configuration.

Track .bin/flush-dns. This handles the flushing of the local system's
DNS.

Track .bin/go-test-cover. This handles the unit testing coverage report
of Go projects.

Track .gitconfig. This is the global Git config for the current user.

Track .github/README.md. This is the README for the GitHub remote.

Track .gitignore. This is the .gitignore for dotfile.

Track .zprofile and .zshrc. Storing this file allows for a familiar ZSH
shell experience without manual configuration.