add GitKraken unsign script
GitKraken (and other signed Electron apps) are subject to performance
issues on macOS Big Sur. This script removes the signature on the
embedded renderer app, making the GitKraken app package a partially
signed package. This script should be run on GitKraken updates until
the issue is resolved on the macOS side.