Add configs for terminal recording
Modify PS1 when ASCIINEMA_REC is set in .zshrc. Username display is
useful in day-to-day terminal operations, but it's not needed for demos.

Set alias for asciinema rec. There is a bug in current asciinema where
it does not set ASCIINEMA_REC environment variable. A workaround for now
is to set it manually.

Disable shell sessions in .zshenv. This removes the "restored sesssion"
message when opening a new terminal for recording.

Set asciinema idle limit to 1 second. Shorten long terminal idles during
recording to 1 second.